plugins: refactor fastapi dependencies
